refactor: repository 패키지 구조 변경- customRepository 패키지 삭제#80
